requirements/base.txt
Django==1.6.2
South==0.8.4
django-extensions==1.3.3
django-taggit==0.11.2
django-tastypie==0.11.0
-e git+https://github.com/felixleong/django-tastypie-extendedmodelresource.git@develop#egg=django-tastypie-extendedmodelresource
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.2.1
six==1.5.2
